backend/requirements.txt
Flask==2.2.2
flask-cors==3.0.10
peewee==3.15.3
psycopg2-binary==2.9.3
pybase64==1.2.3
pydantic==1.10.2
python-dotenv==0.21.0
pyjwt==2.5.0
python-dotenv==0.21.0
pytest==7.2.0
requests==2.28.1